Started fixing issues in GeneticFormFitter. Added getter for points in Form.
Removed default values in
calculcate_best_offset_from_first_to_second_form of GeneticFormFitter.

Added a getter for points in the Form class, in order to being able to
combine the points of multiple forms.

Started added test code for the GeneticFormFitter. Which is running
into a segmentation fault 11 at the moment. This test code is only run,
when the -DDEBUG flag is set in the make file.